+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
 ## [0.6.0-rc.1] - 2025-03-14
 
 ### Added
@@ -155,6 +157,7 @@ _Stable release based on [0.1.0-rc.1]._
 
 ## [0.0.0] - 2024-09-12
 
+[Unreleased]: https://https://github.com/internetguru/laravel-common/compare/staging...dev
 [0.6.0-rc.1]: https://github.com/internetguru/laravel-common/releases/tag/v0.5.4
 [0.5.4]: https://https://github.com/internetguru/laravel-common/compare/v0.5.3...v0.5.4
 [0.5.3]: https://https://github.com/internetguru/laravel-common/compare/v0.5.2...v0.5.3